requirements.txt
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic~=2.9.2
pydantic-settings~=2.6.1
python-dotenv~=1.0.1
Werkzeug~=3.1.1
alembic~=1.13.3
redis~=5.2.0
uvicorn~=0.32.0
fastapi~=0.115.4
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
orjson==3.10.11
PyJWT==2.9.0
werkzeug==3.1.2
pybreaker==1.2.0
passlib==1.7.4
pydantic[email]
uvicorn==0.32.0
gunicorn==23.0.0
aiohttp==3.11.8
python-multipart==0.0.17
async-fastapi-jwt-auth==0.6.6
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2


typing_extensions~=4.12.2
httpx~=0.27.2
starlette~=0.41.3
Django~=5.0.4
elasticsearch~=8.16.0
djangorestframework~=3.15.2
elasticsearch-dsl~=8.16.0
structlog~=24.4.0